+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Fixed riding detection (hopefully temporary).
 * Added sorting.
 * Updated mount display system.
+* Added "Enable Flags" option in map settings.